requirements.txt
transformers
torch
requests
Pillow
fastapi
uvicorn
streamlit
python-multipart